+++ CMakeLists.txt
@@ -136,6 +136,7 @@ set(libsrc
     src/plugins_types.c
     src/plugins_types_binary.c
     src/plugins_types_integer.c
+    src/plugins_types_string.c
     src/plugins_exts.c
     src/plugins_exts_metadata.c
     src/plugins_exts_nacm.c
